This is the fourth subject line
This is the third subject line

This is the subject line of the squashed commit.

This is the subject line.

This is the body of the commit.
There might be more here.

Create pre-push githook

Create pre-push githook

Edit pre-push githook

Rename githooks directory

Edit pre-push git hook

This is the body.
Another line in the body.

Last commit

this is the subject line

This is the subject line of the squashed commit.

This is the subject line.

This is the body of the commit.
There might be more here.

Create pre-push githook

Create pre-push githook

Edit pre-push githook

Rename githooks directory

Edit pre-push git hook

This is the body.
Another line in the body.

Last commit

This is another commit on this branch.

This is the body.
Another line on the body.

This is another commit on this branch.

This is the body.
Another line on the body.

update pre-push hook

add newline

Fail pre-push hook